testclient/requirements.txt
locustio
redis
requests
google-cloud-pubsub